chapter_6/recipe_4/src/requirements.txt
google-cloud-logging
google-cloud-error-reporting
